feat(engine): make blake3 optional
